requirements.txt
Pillow==9.4.0
opencv-python-headless==4.7.0.68
numpy==1.24.1
python-dotenv==0.21.1
Flask==2.2.2
pandas==1.5.3
onnxruntime==1.13.1
